requirements.txt
# Core runtime
python-dotenv>=1.0.0
rich>=13.0.0

# Local LLM backend (optional — requires a .gguf model file)
# Uncomment if you have a local model:
# llama-cpp-python>=0.2.57

# Anthropic Claude backend (optional)
anthropic>=0.40.0

# OpenAI-compatible backend (optional)
openai>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
